Parse #[component(default)] flag
